requirements.txt
pillow==9.4.0
streamlit
gspread
google-auth
